pyproject.toml
[tool.poetry]
name = "rlarcworld"
version = "0.1.0"
description = "An enviroment to train on arc dataset a model-based reinforcement learning agent"
authors = ["atellou <andresfe.t.u@gmail.com>"]
license = "http://www.apache.org/licenses/"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.5.1"
gymnasium = "^1.0.0"
matplotlib = "^3.9.3"
numpy = "^2.2.0"
torchrl = "^0.7.0"
tensorboard = "^2.19.0"
standard-imghdr = "^3.13.0"
torchvision = "^0.21.0"
pyyaml = "^6.0.2"
google-cloud-aiplatform = "^1.81.0"
gcsfs = "^2025.2.0"
google-cloud-logging = "^3.11.4"
python-json-logger = "^3.2.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
